separate callbacks module
