+++ contracts/multisig-improved/src/action_types/mod.rs
@@ -1,3 +1,4 @@
+pub mod callbacks;
 pub mod discard;
 pub mod execute_action;
 pub mod perform;
